+++ llvm/lib/CAS/UnifiedOnDiskCache.cpp
@@ -88,10 +88,6 @@
 #include "llvm/Support/raw_ostream.h"
 #include <optional>
 
-#if __has_include(<sys/sysctl.h>)
-#include <sys/sysctl.h>
-#endif
-
 using namespace llvm;
 using namespace llvm::cas;
 using namespace llvm::cas::ondisk;
@@ -269,29 +265,6 @@ static Error validateInProcess(StringRef RootPath, StringRef HashName,
   if (Error E = Cache->validate())
     return E;
   return Error::success();
-}
-
-static Expected<uint64_t> getBootTime() {
-#if __has_include(<sys/sysctl.h>) && defined(KERN_BOOTTIME)
-  struct timeval TV;
-  size_t TVLen = sizeof(TV);
-  int KernBoot[2] = {CTL_KERN, KERN_BOOTTIME};
-  if (sysctl(KernBoot, 2, &TV, &TVLen, nullptr, 0) < 0)
-    return createStringError(llvm::errnoAsErrorCode(),
-                             "failed to get boottime");
-  if (TVLen != sizeof(TV))
-    return createStringError("sysctl kern.boottime unexpected format");
-  return TV.tv_sec;
-#elif defined(__linux__)
-  // Use the mtime for /proc, which is recreated during system boot.
-  // We could also read /proc/stat and search for 'btime'.
-  sys::fs::file_status Status;
-  if (std::error_code EC = sys::fs::status("/proc", Status))
-    return createFileError("/proc", EC);
-  return Status.getLastModificationTime().time_since_epoch().count();
-#else
-  llvm::report_fatal_error("getBootTime unimplemented");
-#endif
 }
 
 Expected<ValidationResult> UnifiedOnDiskCache::validateIfNeeded(
